+++ test/brando/runtime_config_test.exs
@@ -1,0 +1,20 @@
+defmodule Brando.RuntimeConfigTest do
+  use ExUnit.Case, async: true
+
+  alias Brando.RuntimeConfig
+
+  test "resolves web infrastructure modules without the Brando application facade" do
+    web_module = RuntimeConfig.get(:web_module)
+
+    assert RuntimeConfig.endpoint() == Module.concat(web_module, Endpoint)
+    assert RuntimeConfig.router_helpers() == Module.concat(web_module, Router.Helpers)
+    assert RuntimeConfig.gettext() == Module.concat(web_module, Gettext)
+  end
+
+  test "keeps the Brando application facade compatible" do
+    assert RuntimeConfig.endpoint() == Brando.endpoint()
+    assert RuntimeConfig.router_helpers() == Brando.helpers()
+    assert RuntimeConfig.router_helpers() == Brando.routes()
+    assert RuntimeConfig.gettext() == Brando.gettext()
+  end
+end